always store km in database, display km or mi
Depending on the setting, displayed distance units are either km or mi,
handle neccessary conversion within QML so that data stored in the
database is always in km.

This handling is similar to what fuelpad (on maemo) did and should later
allow more flexibility when displaying various statistics like
consumption etc.

This patch implements this at least for any tank related pages but not
any other information.