Store distances in km, enter in mi (part 2)
Depending on the setting, displayed distance units are either km or mi, handle
neccessary conversion within QML so that data stored in the database is always
in km.

This handling is similar to what fuelpad (on maemo) did and should later allow
more flexibility when displaying various statistics like consumption etc.

This patch implements this for tire information, cost entries and in the budget
overview page.