consumption display defaults to l/100km
Considering that the unit l/100mi is not used at all even in the UK
where miles are still used, it seems to make more sense to display the
consumption in l/100km by default.

Without this commit, the values shown for the consumption is already in
l/100km regardless of distance unit set, so make the label match this.